Always rewind the given IO
When an error occured when extracting dimensions, rewinding would be
skipped and the code would jump straight to "rescue" statements. We want
that the IO is always rewinded, even in case of errors.

Note that this doesn't affect the return value of the method, because
the return value of a method-level "ensure" is always ignored by Ruby.